Move all ID fields to be of type Long.
It looks like the value range can be larger than integer for at least a
few types. As the exact range isn't specified, make sure we're on the
safe side and convert all of them to long.

+++ library/src/main/java/com/meisolsson/githubsdk/model/Deployment.java
@@ -50,7 +50,7 @@ public abstract class Deployment implements Parcelable {
     public abstract String description();
 
     @Nullable
-    public abstract Integer id();
+    public abstract Long id();
 
     @Nullable
     public abstract Map<String, String> payload();
